Function to add matrix .well-known compatibily.
Added needed structs to catch the unmarshaled JSON and the `GetServer`
function.

`GetServer()` follows the Matrix spec by extracting the domain from the user ID to get data about the server the user claims to be trying to connect to. `GetServer()` takes a userID as it's only parameter and returns: The "Home Server", "Identity Server", and an error.

[Link to Matrix spec.](https://matrix.org/docs/spec/client_server/r0.5.0#server-discovery)