Update the Registration struct
Add the `protocols` option. Also add `omitempty` to the `rate_limited` option since it's not required to appear in the registration file, and so we fall back to the homeserver's default value (for example Synapse's is `true` so if one wants to keep to default behaviours they probably don't want it to be automatically set to `false`).