store: Add AccountDataStore
The AccountDataStore leverages a user's account data to load/save the
next_batch token, while retaining the InMemoryStore behaviour for saving
filters and room state.

This permits a client to know which messages it has seen/processed when
(re)syncing, for example after a bot is restarted. This isn't possible
with the InMemoryStore since the token would be lost when the
process is restarted resulting in answering to messages we might have
already handled before.

An otherwise stateless client can now use the home server as its
persistent data store for the next_batch token, instead of having to
use a file or database of its own.