Add option for automatic HTTP request retries on network error
The DefaultHTTPRetries option in Client and AppService sets the number
of retries that *all* HTTP requests do if they encounter a network or
gateway (502-504) error. The default is 0, which means no retries.

It can also be set per-request with FullRequest's MaxAttempts field.
That field starts at 1 instead of 0 so that the default of 0 means
using the default value from the Client.