store: update interface to not take sender key
According to
https://spec.matrix.org/latest/client-server-api/#mmegolmv1aes-sha2,
clients MUST NOT store or lookup sessions using the sender key.

This commit removes the sender key from most of the functions related to
putting and getting group sessions from the Store interface. Notably,
RedactGroupSessions still accepts a sender key because it's meant for
batch deletion of group sessions.

Signed-off-by: Sumner Evans <sumner.evans@automattic.com>

Signed-off-by: Sumner Evans <sumner.evans@automattic.com>