crypto/ed25519: reimplementation with libolm-compatible byte layout
This PR adds a maunium.net/go/mautrix/crypto/ed25519 package to
mautrix-go which is a mirror of the crypto/ed25519 package for
generating Ed25519 signatures, but which uses a different private key
format.

This picture will help with the rest of the explanation:
https://blog.mozilla.org/warner/files/2011/11/key-formats.png

The private key in the [crypto/ed25519] package is a 64-byte value where
the first 32-bytes are the seed and the last 32-bytes are the public
key.

The private key in this package is stored as a 64-byte value that
results from the SHA512 of the seed.

This is the format used by libolm, and is required for pickle/unpickle
to work properly.

Signed-off-by: Sumner Evans <sumner.evans@automattic.com>