Beeper extension for client side media dedup (#267)
The client can send a unique id (like a hash) of a file it intends to
upload in the create request. If the server already has the file it will
return a "completed" response.

The unique id is meant to be opaque for the server. For privacy reasons
it is recommended not to use the raw hash of a file.  The returned MXC
should be stable for the same unique id for the same user but is not
guaranteed.

The room id is used to tie the lifecycle of created media to an existing
room on the homeserver. If a room is purged from the homeserver the
media will be purged along with it.

If the file has been created but not uploaded the response will not have
a completed timestamp which allows the client to retry sending the file.
If the upload has already been completed the upload URL will be empty.

It is possible for multiple clients to send a create request
simultaneously with the same unique id and upload the file at the same
time. It is also possible for the server to forget the unique id and
allow reuploading the same file again returning a new MXC.

This commit also fixes UnusedExpiresAt type in the response which is a
breaking change.