bridgev2/portal: only use event loop when buffer is enabled
When buffer is disabled, queueEvent will instead acquire a lock and call
the handler directly. Hopefully the queueEvent callers are already in a
queue and will block so that queueEvent itself doesn't need to be strictly
FIFO (if callers aren't in a queue, even the buffered channel writes could
race each other).